Modified Transitioning for structure and allocation reduction
Modified a large section of the ITransitionResolutionProvider to provide the foundation for more efficient methods and usage of the provider by the implementer.

The StateMachine now relies on the ITransitionResolutionProvider as per original design and the user may choose to only use the interface or store a reference to a concrete implementation to resolve transitions.